ui: remove tabview
